# BUGFIX: Speed up node tree copy actions 9.0 port of #3984